requirements.txt
# SERVICE-HERO: Automation Platform Dependencies

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Template Engine & Static Files
jinja2==3.1.2
aiofiles==23.2.1

# Database & Caching
sqlalchemy==2.0.23
alembic==1.13.0
redis==5.0.1
databases[postgresql]==0.8.0
asyncpg==0.29.0

# Background Tasks & Queue
celery==5.3.4
kombu==5.3.4

# HTTP Client for Integrations
httpx==0.25.2
requests==2.31.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Communication Integrations
twilio==8.10.3
sendgrid==6.11.0
python-telegram-bot==20.7

# Business Tool Integrations
stripe==7.8.0
hubspot-api-client==7.7.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Scheduling & Calendar
icalendar==5.0.11
pytz==2023.3.post1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Validation & Utilities
email-validator==2.1.0
python-dateutil==2.8.2
pyyaml==6.0.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging  
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Configuration Management
python-dotenv==1.0.0
pydantic-settings==2.1.0

# API Documentation
markdown==3.5.1

# File Handling
openpyxl==3.1.2
python-magic==0.4.27

# WebSocket Support (for real-time features)
websockets==12.0

# Rate Limiting
slowapi==0.1.9

# CORS Support
fastapi-cors==0.0.6

# Production Server
gunicorn==21.2.0